requirements.txt
appdirs==1.4.4 ; python_version >= "3.10" and python_version < "3.12"
art==6.2 ; python_version >= "3.10" and python_version < "3.12"
beartype==0.17.2 ; python_version >= "3.10" and python_version < "3.12"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.12" and (platform_system == "Windows" or sys_platform == "win32")
contourpy==1.2.1 ; python_version >= "3.10" and python_version < "3.12"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "3.12"
docker-pycreds==0.4.0 ; python_version >= "3.10" and python_version < "3.12"
einops==0.7.0 ; python_version >= "3.10" and python_version < "3.12"
executing==2.0.1 ; python_version >= "3.10" and python_version < "3.12"
eztils==0.4.94 ; python_version >= "3.10" and python_version < "3.12"
fonttools==4.51.0 ; python_version >= "3.10" and python_version < "3.12"
gitdb==4.0.11 ; python_version >= "3.10" and python_version < "3.12"
gitpython==3.1.43 ; python_version >= "3.10" and python_version < "3.12"
idna==3.7 ; python_version >= "3.10" and python_version < "3.12"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "3.12"
loguru==0.7.2 ; python_version >= "3.10" and python_version < "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "3.12"
matplotlib==3.8.4 ; python_version >= "3.10" and python_version < "3.12"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "3.12"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "3.12"
packaging==24.0 ; python_version >= "3.10" and python_version < "3.12"
pandas==2.0.3 ; python_version >= "3.10" and python_version < "3.12"
pillow==10.3.0 ; python_version >= "3.10" and python_version < "3.12"
protobuf==4.25.3 ; python_version >= "3.10" and python_version < "3.12"
psutil==5.9.8 ; python_version >= "3.10" and python_version < "3.12"
pygments==2.17.2 ; python_version >= "3.10" and python_version < "3.12"
pyparsing==3.1.2 ; python_version >= "3.10" and python_version < "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "3.12"
pytz==2024.1 ; python_version >= "3.10" and python_version < "3.12"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.12"
rich==13.7.1 ; python_version >= "3.10" and python_version < "3.12"
seaborn==0.13.2 ; python_version >= "3.10" and python_version < "3.12"
sentry-sdk==2.0.1 ; python_version >= "3.10" and python_version < "3.12"
setproctitle==1.3.3 ; python_version >= "3.10" and python_version < "3.12"
setuptools==69.5.1 ; python_version >= "3.10" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.12"
smmap==5.0.1 ; python_version >= "3.10" and python_version < "3.12"
tqdm==4.66.2 ; python_version >= "3.10" and python_version < "3.12"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "3.12"
urllib3==2.2.1 ; python_version >= "3.10" and python_version < "3.12"
varname==0.13.0 ; python_version >= "3.10" and python_version < "3.12"
wandb==0.16.6 ; python_version >= "3.10" and python_version < "3.12"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "win32"
